+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Added
 - More images to shop/beach
+- Screen to choose which beach save slot to use
+- Buy 2nd and 3rd beaches with 3000 sand dollars
 
 ### Changed
 - When clicking an item on the beach, it will automatically be layered to the top